org-mode style output; added opennlpAggressive tokenizer; models from config
The pre-proc_tests shell script now output in a format more
compatible with org-mode style printing.

The OpenNLP tokenizer was split into a vanilla version and Young's
aggressive rules built on top of the vanilla tokenizer.

Models for the OpenNLP tokenizer are loaded from a path
provided as part of a config (rather than hard-coded).