input and output base dir as options; new sentence splitter option
The base directory for both input and output files can now be specified
in the command line.  Otherwise, they default to as before. The
output is just the root dir because the leaf directory is determined
by the system breadcrumbs so that each run gets its own subfolder.

The base OpenNLP sentence splitter was augmented in parallel with the
tokenizer so that you can run the base model or also run some
additional post-processing on the output f the base model. For now,
there is one patch option that splits in long stretches of whitespace
(3+ in a row). These help make RDW notes better (but still not perfect).