added support for testing two cTAKES sentence splitters
The default cTAKES SentenceDetector and SentenceDetectorBIO
were added as new options (called ctakes and ctakesBIO in terms
of command line options and output breadcrumbs). We needed to add
a new SourceDocumentInformation type and make some updates
to the pre-processing shell script. Both sentence splitters
require a pre-annotated segment to split within so we use
cTAKES simple segmenter that puts the whole document into a single
large segment. Model files need to be copied over from the cTAKES
directory (as with the OpenNLP models).